requirements.txt
flask
streamlit
requests
langchain
pandas
python-docx
python-pptx
PyPDF2
sqlite3
sentence-transformers
faiss-cpu
